submit form

+++ pages/forms.py
@@ -2,4 +2,4 @@ from django import forms
 
 class PagesForm(forms.Form):
     item_name = forms.CharField(label = 'Item Name', max_length = 120)
-    zip_code = forms.DecimalField(label = 'Zip Code')
+    zip_code = forms.IntegerField(label = 'Zip Code')
